pdf-field-locator/requirements.txt
flask
flask_cors
pdfminer.six
simplejson
raven[flask]
